Revert "feat(zip): extract one file (#2603)"
This reverts commit b63db2be7d7138881bc4d07b3d060b522dc09158.
