--- modules/zip/src/types.ts
@@ -1,1 +1,0 @@
-export type FileMap = Record<string, ArrayBuffer>;
