+++ modules/zip/test/zip-writer-loader.spec.js
@@ -2,9 +2,7 @@ import test from 'tape-promise/tape';
 import {validateLoader, validateWriter} from 'test/common/conformance';
 
 import {ZipWriter, ZipLoader} from '@loaders.gl/zip';
-import {encode, load, parse} from '@loaders.gl/core';
-import {getFileProvider} from './test-utils/get-file-provider';
-import {ZipFileSystem} from '../src/filesystems/zip-filesystem';
+import {encode, parse} from '@loaders.gl/core';
 
 const FILE_MAP = {
   src: 'abc',
@@ -12,8 +10,6 @@ const FILE_MAP = {
   'README.md': 'This is a module',
   package: '{"name": "module"}'
 };
-
-const ZIP_FILE_PATH = './modules/zip/test/data/test-store.zip';
 
 test('Zip#loader/writer conformance', (t) => {
   validateLoader(t, ZipLoader, 'ZipLoader');
@@ -35,19 +31,4 @@ test('Zip#encode/decode', (t) => {
       t.fail(error.message);
       t.end();
     });
-});
-
-test('Zip#load one file', async (t) => {
-  const fileProvider = await getFileProvider(ZIP_FILE_PATH);
-  const fileSystem = new ZipFileSystem(fileProvider);
-  const result = await load('test-file.txt', ZipLoader, {
-    fetch: fileSystem.fetch.bind(fileSystem),
-    zip: {mode: 'one'}
-  });
-  if (result instanceof ArrayBuffer) {
-    const text = new TextDecoder().decode(result);
-    t.equal(text, 'test file data\n');
-  }
-  fileSystem.destroy();
-  t.end();
 });
